Resetting the whole thing
Broken layout in past commits. Clearing away all layout and starting from scratch. This is just basic navigation for a style guide, but also leads to a bare-bones main page. starting the page structures. The plan going forward is to fill in my style guide with the components i'll be using throughout the main thing, re-familiarize and orient myself with the project.